Add adventure files from ifarchive.org
Files extracted from:
 - https://www.ifarchive.org/if-archive/scott-adams/games/scottfree/AdamsGames.zip
 - https://www.ifarchive.org/if-archive/scott-adams/games/scottfree/desert-scott.zip
 - https://www.ifarchive.org/if-archive/scott-adams/games/scottfree/ghostking.zip
 - https://www.ifarchive.org/if-archive/scott-adams/games/scottfree/mysterious.tar.gz
 - https://www.ifarchive.org/if-archive/scott-adams/games/scottfree/otheradv.zip

Some README files have been renamed and source code omitted.

I have permission from Scott Adams to include his adventures.
The Brian Howarth adventures are "supposedly" in the Public Domain.
The copyright status of some of the miscellaneous adventures is unknown.

If you are the copyright holder of any of this content and do not want
it to be distributed please get in contact so I can remove it.